Move smallAnnotation up
