+++ src/main/java/com/brunodles/jsoupparser/collectors/AttrCollectorTransformer.java
@@ -1,8 +1,8 @@
-package com.brunodles.jsoupparser.smallanotation.collectors;
+package com.brunodles.jsoupparser.collectors;
 
 import com.brunodles.jsoupparser.Transformer;
-import com.brunodles.jsoupparser.smallanotation.AnnotationInvocation;
-import com.brunodles.jsoupparser.smallanotation.TransformerFor;
+import com.brunodles.jsoupparser.AnnotationInvocation;
+import com.brunodles.jsoupparser.TransformerFor;
 import org.jsoup.nodes.Element;
 
 @TransformerFor(AttrCollector.class)
