+++ src/main/java/com/brunodles/jsoupparser/navigate/Navigate.java
@@ -1,4 +1,4 @@
-package com.brunodles.jsoupparser.smallanotation.navigate;
+package com.brunodles.jsoupparser.navigate;
 
 import java.lang.annotation.*;
