+++ src/main/java/com/brunodles/jsoupparser/navigate/NavigateTransformer.java
@@ -1,8 +1,8 @@
-package com.brunodles.jsoupparser.smallanotation.navigate;
+package com.brunodles.jsoupparser.navigate;
 
 import com.brunodles.jsoupparser.Transformer;
-import com.brunodles.jsoupparser.smallanotation.AnnotationInvocation;
-import com.brunodles.jsoupparser.smallanotation.TransformerFor;
+import com.brunodles.jsoupparser.AnnotationInvocation;
+import com.brunodles.jsoupparser.TransformerFor;
 
 import java.util.ArrayList;
 import java.util.List;
