+++ src/main/java/com/brunodles/jsoupparser/selector/Selector.java
@@ -1,10 +1,10 @@
-package com.brunodles.jsoupparser.smallanotation.collectors;
+package com.brunodles.jsoupparser.selector;
 
 import java.lang.annotation.*;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface AttrCollector {
+public @interface Selector {
     String value();
 }
