+++ src/test/java/com/brunodles/jsoupparser/SmallAnnotation.java
@@ -1,13 +1,12 @@
 package com.brunodles.jsoupparser;
 
 import com.brunodles.jsoupparser.transformers.TransformToFloat;
-import com.brunodles.jsoupparser.smallanotation.SmallAnnotationInvocationHandler;
-import com.brunodles.jsoupparser.smallanotation.collectors.AttrCollector;
-import com.brunodles.jsoupparser.smallanotation.navigate.Navigate;
-import com.brunodles.jsoupparser.smallanotation.nested.Nested;
-import com.brunodles.jsoupparser.smallanotation.withtype.WithType;
-import com.brunodles.jsoupparser.smallanotation.collectors.TextCollector;
-import com.brunodles.jsoupparser.smallanotation.selector.Selector;
+import com.brunodles.jsoupparser.collectors.AttrCollector;
+import com.brunodles.jsoupparser.navigate.Navigate;
+import com.brunodles.jsoupparser.nested.Nested;
+import com.brunodles.jsoupparser.withtype.WithType;
+import com.brunodles.jsoupparser.collectors.TextCollector;
+import com.brunodles.jsoupparser.selector.Selector;
 import com.brunodles.test.ResourceUriResolver;
 import org.junit.Test;
 import org.junit.runner.RunWith;
