+++ src/test/java/com/brunodles/jsoupparser/WhenHaveError.java
@@ -3,8 +3,8 @@ package com.brunodles.jsoupparser;
 import com.brunodles.jsoupparser.exceptions.InvalidResultException;
 import com.brunodles.jsoupparser.exceptions.InvalidSelectorException;
 import com.brunodles.jsoupparser.exceptions.MissingSelectorException;
-import com.brunodles.jsoupparser.smallanotation.collectors.TextCollector;
-import com.brunodles.jsoupparser.smallanotation.selector.Selector;
+import com.brunodles.jsoupparser.collectors.TextCollector;
+import com.brunodles.jsoupparser.selector.Selector;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
