+++ src/test/java/com/brunodles/jsoupparser/WhenParseNestedContent.java
@@ -1,8 +1,8 @@
 package com.brunodles.jsoupparser;
 
-import com.brunodles.jsoupparser.smallanotation.collectors.TextCollector;
-import com.brunodles.jsoupparser.smallanotation.nested.Nested;
-import com.brunodles.jsoupparser.smallanotation.selector.Selector;
+import com.brunodles.jsoupparser.collectors.TextCollector;
+import com.brunodles.jsoupparser.nested.Nested;
+import com.brunodles.jsoupparser.selector.Selector;
 import org.junit.Before;
 import org.junit.Test;
